fixed amadec ioctl bug on 64bit linux.

+++ package/multimedia/aml_libs/src/amadec/audio-dec.h
@@ -35,16 +35,15 @@ ADEC_BEGIN_DECLS
 
 //for ffmpeg audio decode
 #define AMSTREAM_IOC_MAGIC  'S'
-#define AMSTREAM_IOC_APTS_LOOKUP    _IOR(AMSTREAM_IOC_MAGIC, 0x81,unsigned long)   
-#define GET_FIRST_APTS_FLAG			_IOR(AMSTREAM_IOC_MAGIC, 0x82, long)
-
+#define AMSTREAM_IOC_APTS_LOOKUP    _IOR((AMSTREAM_IOC_MAGIC), 0x81, int)  
+#define GET_FIRST_APTS_FLAG    _IOR((AMSTREAM_IOC_MAGIC), 0x82, int)
 //-----------------------------------------------
 //copy from file: "../amcodec/include/amports/amstream.h"
 #ifndef AMSTREAM_IOC_PCRSCR
-#define AMSTREAM_IOC_PCRSCR           _IOR(AMSTREAM_IOC_MAGIC, 0x42, unsigned long)
+#define AMSTREAM_IOC_PCRSCR  _IOR((AMSTREAM_IOC_MAGIC), 0x42, int)
 #endif
 #ifndef AMSTREAM_IOC_SET_APTS
-#define AMSTREAM_IOC_SET_APTS         _IOW(AMSTREAM_IOC_MAGIC, 0xa8, unsigned long)
+#define AMSTREAM_IOC_SET_APTS  _IOW((AMSTREAM_IOC_MAGIC), 0xa8, int)
 #endif
 
 //-----------------------------------------------
